// dashboard/src/app/core/models/worker.models.ts
// worker.model.ts
import {WorkerSessionModels} from "./worker-session.models";

export class WorkerModel {
    id: number;
    email: string;
    firstName: string;
    lastName: string;
    password:string;
    role: string;
    workerSessions:WorkerSessionModels[];
    createdAt: string;
    updatedAt: string;

}